src/main/java/jpabook/jpashop/domain/MemberRepository.java
package jpabook.jpashop.domain;

import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

@Repository
public class MemberRepository {

    @PersistenceContext
    private EntityManager em;

    public long save(Member member){
        em.persist(member);
        return member.getId();
    }

    public Member find(long id){
        return em.find(Member.class,id);
    }
}
